apps/backend/requirements.txt
# FastAPI и зависимости
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.1.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
bcrypt==4.1.3

# HTTP клиент
httpx==0.28.1
aiohttp==3.11.10

# CORS
python-dotenv==1.0.1

# Опциональные зависимости для production
# gunicorn==23.0.0
# sentry-sdk[fastapi]==2.19.2
# redis==5.2.0
# celery==5.4.0

# Development dependencies (устанавливаются отдельно в dev)
# pytest==8.3.4
# pytest-asyncio==0.24.0
# pytest-cov==6.0.0
# black==24.10.0
# flake8==7.1.1
# mypy==1.13.0
